% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{chain_investment}
\alias{chain_investment}
\title{BEA Infrastructure Investment}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6035 rows and 5 columns.
}
\usage{
chain_investment
}
\description{
BEA Infrastructure Investment
}
\references{
\url{https://github.com/rfordatascience/tidytuesday/blob/master/data/2021/2021-08-10/readme.md}
}
\author{
tidytuesday
}
\keyword{datasets}
