% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{datasets}
\alias{datasets}
\title{BEA Infrastructure Investment}
\format{
An object of class \code{list} of length 3.
}
\usage{
datasets
}
\description{
A list containing three datasets.
}
\references{
\url{https://github.com/rfordatascience/tidytuesday/blob/master/data/2021/2021-08-10/readme.md}
}
\author{
tidytuesday
}
\keyword{datasets}
