% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_category.R
\name{plt_category}
\alias{plt_category}
\title{Plot trends grouped by category}
\usage{
plt_category(df)
}
\arguments{
\item{df}{dataframe containing bea variables}
}
\description{
Plot trends grouped by category
}
