% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_category.R
\name{plt_category_pop}
\alias{plt_category_pop}
\title{Plot trends grouped by category and choose the y axis variable.}
\usage{
plt_category_pop(df, var_name)
}
\arguments{
\item{df}{dataframe containing bea variables}

\item{var_name}{either gross_inv_per_cap or gross_inv_chain}
}
\description{
Plot trends grouped by category and choose the y axis variable.
}
