% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{us_population_1960_2017}
\alias{us_population_1960_2017}
\title{U.S total population by year}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58 rows and 3 columns.
}
\usage{
us_population_1960_2017
}
\description{
U.S total population by year
}
\references{
\url{https://data.worldbank.org/indicator/SP.POP.TOTL?locations=US}
}
\author{
Juan Dugo \email{juanpatriciodugo@gmail.com}
}
\keyword{datasets}
